Extract and export the `set` serializer
