+++ packages/umi-serializer-data-view/test/set.test.ts
@@ -1,10 +1,11 @@
 import test from 'ava';
 import { Endian } from '@metaplex-foundation/umi';
+import { set } from '../src/set';
 import { createDataViewSerializer } from '../src';
 import { s, d } from './_helpers';
 
 test('prefixed (de)serialization', (t) => {
-  const { set, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, set(u8()), new Set(), '00000000'); // 4-bytes prefix.
@@ -30,7 +31,7 @@ test('prefixed (de)serialization', (t) => {
 });
 
 test('fixed (de)serialization', (t) => {
-  const { set, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, set(u8(), { size: 0 }), new Set(), '');
@@ -71,7 +72,7 @@ test('fixed (de)serialization', (t) => {
 });
 
 test('remainder (de)serialization', (t) => {
-  const { set, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
   const remainder = { size: 'remainder' } as const;
 
   // Empty.
@@ -100,7 +101,7 @@ test('remainder (de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { set, u8, u16 } = createDataViewSerializer();
+  const { u8, u16 } = createDataViewSerializer();
 
   // Size.
   t.is(set(u8(), { size: 42 }).description, 'set(u8; 42)');
@@ -120,7 +121,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { set, u8, u32, string } = createDataViewSerializer();
+  const { u8, u32, string } = createDataViewSerializer();
   t.is(set(u8()).fixedSize, null);
   t.is(set(u8()).maxSize, null);
   t.is(set(u8(), { size: u8() }).fixedSize, null);
